requirements.txt
accelerate==0.31.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
backgroundremover==0.2.8
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
commandlines==0.4.1
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
diffusers==0.29.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.15.4
filetype==1.2.0
fonttools==4.53.0
fsspec==2024.6.0
future==1.0.0
gradio==4.36.1
gradio_client==1.0.1
h11==0.14.0
hsh==1.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==7.2.1
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
opencv-python==4.10.0.84
orjson==3.10.5
packaging==24.1
pandas==2.2.2
Pillow==9.5.0
proglog==0.1.10
psutil==6.0.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.10
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
semantic-version==2.10.0
setuptools==70.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
tifffile==2024.6.18
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
waitress==3.0.0
watchfiles==0.22.0
websockets==11.0.3
zipp==3.19.2
